requirements.txt
asgiref==3.8.1
backports-zoneinfo==0.2.1
black==24.8.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
django==4.2.24
djangorestframework==3.15.2
idna==3.10
mypy-extensions==1.1.0
packaging==25.0
paho-mqtt==2.1.0
pathspec==0.12.1
platformdirs==4.3.6
requests==2.32.4
sqlparse==0.5.3
tomli==2.2.1
typing-extensions==4.13.2
urllib3==2.2.3
asgiref==3.8.1
backports-zoneinfo==0.2.1
black==24.8.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
django==4.2.24
djangorestframework==3.15.2
dotenv==0.9.9
idna==3.10
mypy-extensions==1.1.0
packaging==25.0
paho-mqtt==2.1.0
pathspec==0.12.1
platformdirs==4.3.6
python-dotenv==1.0.1
requests==2.32.4
sqlparse==0.5.3
tomli==2.2.1
typing-extensions==4.13.2
urllib3==2.2.3
